+++ win-linux/src/ceditorwindow.h
@@ -84,9 +84,9 @@ protected:
     void onSizeEvent(int) override;
     void onMoveEvent(const QRect&) override;
     void onExitSizeMove() override;
-    void onDpiChanged(int,int) override;
+    void onDpiChanged(double,double) override;
 
-    void setScreenScalingFactor(int) override;
+    void setScreenScalingFactor(double) override;
     int calcTitleCaptionWidth() override;
 
 private slots:
